+++ ui-events/CHANGELOG.md
@@ -15,11 +15,19 @@ You can find its changes [documented below](#010-2025-05-08).
 
 This release has an [MSRV][] of 1.73.
 
+### Added
+
+* `PointerId`, `PointerInfo`, `PointerUpdate`, and `PointerEvent` now have a `is_primary_pointer` methhod. ([#54][] by [@waywardmonkeys][])
+
 ## [0.1.0][] - 2025-05-08
 
 This release has an [MSRV][] of 1.73.
 
 This is the initial release.
+
+[@waywardmonkeys]: https://github.com/waywardmonkeys
+
+[#54]: https://github.com/endoli/ui-events/pull/54
 
 [Unreleased]: https://github.com/endoli/ui-events/compare/v0.1.0...HEAD
 [0.1.0]: https://github.com/linebender/color/releases/tag/v0.1.0